Refactor App: remove redux
