+++ src/App.jsx
@@ -1,42 +1,17 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
-import { showHeader, hideHeader } from './slices/headerSlice';
+import React from 'react';
 
 import Footer from './components/Footer';
 import Header from './components/header/Header';
 import HeaderFixed from './components/header-fixed/HeaderFixed';
 import Main from './components/Main';
 
-const App = () => {
-  const dispatch = useDispatch();
-  const isShow = useSelector((state) => state.header.isShow);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const distance = 450;
-      if (window.scrollY >= distance && !isShow) {
-        dispatch(showHeader());
-      } else if (window.scrollY < distance && isShow) {
-        dispatch(hideHeader());
-      }
-    };
-
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [isShow, dispatch]);
-
-  return (
-    <>
-      <Header />
-      <HeaderFixed />
-      <Main />
-      <Footer />
-    </>
-  );
-};
+const App = () => (
+  <>
+    <Header />
+    <HeaderFixed />
+    <Main />
+    <Footer />
+  </>
+);
 
 export default App;
