--- src/slices/headerSlice.js
@@ -1,23 +1,0 @@
-/* eslint no-param-reassign: ["error", { "props": false }] */
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  isShow: false,
-};
-
-const headerSlice = createSlice({
-  name: 'header',
-  initialState,
-  reducers: {
-    showHeader: (state) => {
-      state.isShow = true;
-    },
-    hideHeader: (state) => {
-      state.isShow = false;
-    },
-  },
-});
-
-export const { showHeader, hideHeader } = headerSlice.actions;
-
-export default headerSlice.reducer;
