--- src/slices/index.js
@@ -1,9 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-
-import headerReducer from './headerSlice.js';
-
-export default configureStore({
-  reducer: {
-    header: headerReducer,
-  },
-});
